+++ tests/testthat/test-tar_quarto_files.R
@@ -16,7 +16,7 @@ targets::tar_test("tar_quarto_files() single Rmd/qmd", {
     out <- tar_quarto_files(path)
     expect_equal(out$sources, file.path("x", paste0("report", ext)))
     expect_equal(out$output, file.path("x", "report.html"))
-    expect_equal(out$input, file.path("x", paste0("report", ext)))
+    expect_equal(out$input, character(0))
   }
 })
 
@@ -62,55 +62,49 @@ targets::tar_test("tar_quarto_files() project", {
       sort(c("index.qmd", "r2.qmd"))
     )
     expect_equal(basename(info$output), "_book")
-    expect_equal(
-      sort(basename(info$input)),
-      sort(c("_quarto.yml", "index.qmd", "r2.qmd"))
-    )
+    expect_equal(basename(info$input), "_quarto.yml")
   } else {
     expect_equal(
       sort(info$sources),
       sort(file.path("x", c("index.qmd", "r2.qmd")))
     )
     expect_equal(info$output, file.path("x", "_book"))
-    expect_equal(
-      sort(info$input),
-      sort(file.path("x", c("_quarto.yml", "index.qmd", "r2.qmd")))
-    )
+    expect_equal(info$input, file.path("x", "_quarto.yml"))
   }
 })
 
-targets::tar_test("tar_quarto_files() detects non-code dependencies", {
+targets::tar_test("tar_quarto_files() detects nested dependencies", {
   skip_quarto()
   fs::dir_create("report")
   fs::dir_create(file.path("report", "subdir"))
+  fs::dir_create(file.path("report", "subdir", "b"))
   lines <- c(
     "---",
     "title: main",
     "output_format: html",
     "---",
     "",
-    "{{< include \"text1.qmd\" >}}",
-    "",
-    "{{< include \"subdir/text2.qmd\" >}}"
+    "{{< include \"text1.qmd\" >}}"
   )
   writeLines(lines, file.path("report", "main.qmd"))
   lines <- c(
     "# First File",
     "",
-    "Some text here."
+    "Some text here.",
+    "",
+    "{{< include \"subdir/text2.qmd\" >}}"
   )
   writeLines(lines, file.path("report", "text1.qmd"))
   lines <- c(
     "# Second File",
     "",
-    "Some text here."
+    "Some text here.",
+    ""
   )
   writeLines(lines, file.path("report", "subdir", "text2.qmd"))
-  out <- tar_quarto_files("report/main.qmd")
-  expect_equal(out$sources, "report/main.qmd")
-  expect_equal(out$output, "report/main.html")
+  out <- tar_quarto_files(file.path("report", "main.qmd"))
   expect_equal(
-    sort(out$input),
+    sort(out$sources),
     sort(
       c(
         file.path("report", c("main.qmd", "text1.qmd")),
@@ -118,8 +112,10 @@ targets::tar_test("tar_quarto_files() detects non-code dependencies", {
       )
     )
   )
+  expect_equal(out$output, file.path("report", "main.html"))
+  expect_equal(out$input, character())
   # Check whether we get the same result with a reference to the directory
-  # instead of the file.
+  # instead of the file (render a project).
   lines <- c(
     "project:",
     "  output-dir: myoutdir",
@@ -129,26 +125,49 @@ targets::tar_test("tar_quarto_files() detects non-code dependencies", {
   writeLines(lines, file.path("report", "_quarto.yml"))
   out <- tar_quarto_files("report/")
   if (identical(tolower(Sys.info()[["sysname"]]), "windows")) {
-    expect_equal(basename(out$sources), "main.qmd")
-    expect_equal(basename(out$output), "myoutdir")
     expect_equal(
-      sort(basename(out$input)),
-      sort(c("_quarto.yml", "main.qmd", "text1.qmd", "text2.qmd"))
+      basename(out$sources),
+      sort(
+        c("main.qmd", "text1.qmd", "text2.qmd")
+      )
     )
+    expect_equal(basename(out$output), "myoutdir")
+    expect_equal(basename(out$input), "_quarto.yml")
   } else {
-    expect_equal(out$sources, file.path("report", "main.qmd"))
-    expect_equal(out$output, file.path("report", "myoutdir"))
     expect_equal(
-      sort(out$input),
+      out$sources,
       sort(
         c(
-          file.path(
-            "report",
-            c("_quarto.yml", "main.qmd", "text1.qmd")
-          ),
+          file.path("report", c("main.qmd", "text1.qmd")),
           file.path("report", "subdir", "text2.qmd")
         )
       )
     )
+    expect_equal(out$output, file.path("report", "myoutdir"))
+    expect_equal(
+      out$input,
+      file.path("report", "_quarto.yml")
+    )
   }
+})
+
+targets::tar_test("tar_quarto_files() detects custom output file", {
+  skip_quarto()
+  lines <- c(
+    "---",
+    "title: source file",
+    "format:",
+    "  html:",
+    "    output-file: custom.html",
+    "---",
+    "Assume these lines are in report.qmd.",
+    "```{r}",
+    "1 + 1",
+    "```"
+  )
+  writeLines(lines, "report.qmd")
+  out <- tar_quarto_files("report.qmd")
+  expect_equal(out$sources, "report.qmd")
+  expect_equal(out$output, "custom.html")
+  expect_equal(out$input, character(0))
 })